Moved the declaration of built-in COM interfaces into separate file
Declaring the com_interface macro with a double colon caused the following
compiler error: "error[E0658]: paths of length greater than one in
macro invocations are currently unstable"
